+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All breaking changes are marked with [BC] and potentially require API consumer changes after updating to the respective version.
+## UNRELEASED
+### Added
+- Add `LineEdit::replace` as an alternative to `LineEdit::set` tries not to change the cursor position.
+- Add `LineEdit::cursor_pos` to retrieve (byte) cursor position .
+- Implement `Deref<Target=LineEdit>` for PromptLine
+### Fixed
+- Fix erasing characters in `LineEdit`.
 
 ## [0.2.1] - 2019-07-21
 ### Fixed